+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.1]
+- Block Start button until experiment is loaded
+
 ## [1.6.0] - 2020-07-06
 - Added a list of created experiments
 - Fixed some bugs with cookies and experiment resume